📦 NEW: set up chakra theme dirs
