+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
 import { ChakraProvider } from '@chakra-ui/react';
-import customTheme from 'styles/theme';
+import customTheme from 'theme';
 
 import 'styles/index.css';
