+++ styles/index.css
@@ -1,6 +1,9 @@
 @import 'nprogress.css';
 @import 'prism.css';
 
+* {
+  font-feature-settings: 'cv02', 'cv03', 'cv04', 'cv11';
+}
 
 .text-gradient {
   background: radial-gradient(
