+++ theme/components/button.ts
@@ -1,0 +1,9 @@
+const Button = {
+  baseStyle: {
+    _focus: {
+      boxShadow: 'none'
+    }
+  }
+};
+
+export default Button;
