+++ theme/components/index.ts
@@ -1,0 +1,11 @@
+import Button from './button';
+import Card from './card';
+import Heading from './heading';
+import Link from './link';
+
+export default {
+  Button,
+  Card,
+  Heading,
+  Link
+};
