+++ theme/components/link.ts
@@ -1,0 +1,9 @@
+const Link = {
+  baseStyle: {
+    _hover: {
+      textDecoration: 'none'
+    }
+  }
+};
+
+export default Link;
