+++ theme/foundations/index.ts
@@ -1,0 +1,7 @@
+import colors from './colors';
+import typography from './typography';
+
+export default {
+  colors,
+  ...typography
+};
