+++ theme/foundations/typography.ts
@@ -1,0 +1,9 @@
+const typography = {
+  fonts: {
+    heading: `Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"`,
+    body: `Inter,-apple-system,BlinkMacSystemFont,"Segoe UI",Helvetica,Arial,sans-serif,"Apple Color Emoji","Segoe UI Emoji","Segoe UI Symbol"`,
+    mono: `"Roboto Mono",Menlo,Monaco,Consolas,"Liberation Mono","Courier New",monospace;`
+  }
+};
+
+export default typography;
